+++ packages/components/src/components/radio-button-group/radio-button-group.tsx
@@ -1,6 +1,7 @@
 import { Component, Element, h, Prop } from '@stencil/core';
 import classNames from 'classnames';
 import statusNote from '../../utils/status-note';
+import { generateUniqueId } from '../../utils/utils';
 
 @Component({
   tag: 'scale-radio-button-group',
@@ -17,6 +18,8 @@ export class RadioButtonGroup {
   @Prop() status?: string = '';
   /** (optional) Input status */
   @Prop() invalid?: boolean = false;
+
+  private readonly internalId = generateUniqueId();
 
   componentDidRender() {
     if (this.status !== '') {
@@ -42,6 +45,7 @@ export class RadioButtonGroup {
   }
 
   render() {
+    const helperTextId = `helper-message-${this.internalId}`;
     return (
       <fieldset class="radio-button-group">
         <legend class="radio-button-group__title">
@@ -54,6 +58,7 @@ export class RadioButtonGroup {
           {this.helperText ? (
             <div
               role="text"
+              id={helperTextId}
               class={this.getCssClassMap()}
               aria-label={this.helperText}
             >