+++ packages/storybook-vue/stories/components/beta-brand-header-navigation/SlotLangSwitcher.vue
@@ -1,0 +1,13 @@
+<template>
+  <scale-telekom-nav-list variant="lang-switcher" slot="lang-switcher" alignment="right">
+    <scale-telekom-nav-item active>
+      <a href="#">EN</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item>
+      <a href="#">SO</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item>
+      <a href="#">DE</a>
+    </scale-telekom-nav-item>
+  </scale-telekom-nav-list>
+</template>